+++ resources/views/app.blade.php
@@ -4,7 +4,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title inertia>Vyzor - Inertia + Vue & Laravel Admin & Dashboard Template</title>
+  <title inertia>{{ app(\App\Settings\GeneralSettings::class)->app_name }}</title>
   <!-- Favicon -->
   <link rel="icon" href="{{asset('/images/brand-logos/favicon.ico')}}" type="image/x-icon">
   <!-- Inter Font -->
@@ -17,7 +17,7 @@
   @vite(['resources/js/app.js'])
   @inertiaHead
 
-  <meta name="description" content="Vyzor Admin Dashboard Template built with Vue.js and Laravel. Easily manage your admin panel with our custom template." />
+  <meta name="description" content="{{ app(\App\Settings\GeneralSettings::class)->app_description }}" />
   <meta name="keywords" content="admin dashboard laravel,admin panel laravel,admin panel template,admin template for laravel,dashboard admin template,dashboard ui design,inertia js vue,laravel dashboard,laravel inertia js,laravel template admin,laravel user interface,vue dashboard template,vue laravel admin template,vue ui framework,vuejs admin panel" />
 
 </head>